Split wu integration tests from pure unittests
